+++ back/controllers/ctrlPostLike.js
@@ -12,7 +12,7 @@ exports.getAllPostLikes = async (req, res, next) => {
 
 exports.createPostLike = async (req, res, next) => {
     try {
-        const newPostLike = { ...req.body };
+        const newPostLike = { ...req.body, user_id: req.auth.user_id};
         const postLike = await PostLike.create(newPostLike);
         res.status(200).json(postLike);
     } catch (error) {
@@ -40,6 +40,9 @@ exports.deletePostLike = async (req, res, next) => {
         const postLike = await PostLike.findByPk(postLikeId); 
         if (!postLike) {
             return res.status(404).json({ message: 'postLike not found!' }); 
+        }
+        if (postLike.user_id !== req.auth.user_id) {
+            return res.status(403).json({ message: 'Forbidden' });
         }
         await PostLike.destroy({where: { id: postLikeId }});
         res.status(200).json({ message: 'postLike deleted!' });
